Add server-side processing for datatables, vue when creating a lens, and authentication.

+++ app/Http/Controllers/LensController.php
@@ -1,20 +1,95 @@
 <?php
+/**
+ * Lens Controller.
+ *
+ * PHP Version 7
+ *
+ * @category Lenses
+ * @package  DeepskyLog
+ * @author   Wim De Meester <deepskywim@gmail.com>
+ * @license  GPL3 <https://opensource.org/licenses/GPL-3.0>
+ * @link     http://www.deepskylog.org
+ */
 
 namespace App\Http\Controllers;
 
 use App\Lens;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use App\DataTables\LensDataTable;
 
+/**
+ * Lens Controller.
+ *
+ * @category Lenses
+ * @package  DeepskyLog
+ * @author   Wim De Meester <deepskywim@gmail.com>
+ * @license  GPL3 <https://opensource.org/licenses/GPL-3.0>
+ * @link     http://www.deepskylog.org
+ */
 class LensController extends Controller
 {
+    /**
+     * Only make sure the lens pages can be seen if the user is authenticated
+     * and verified.
+     */
+    public function __construct()
+    {
+        $this->middleware(['auth', 'verified'])->except(['show']);
+    }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(LensDataTable $dataTable)
     {
-        return view('layout.lens.view');
+        return $this->_indexView($dataTable, "user");
+    }
+
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function indexAdmin(LensDataTable $dataTable)
+    {
+        if (auth()->user()->isAdmin()) {
+            //$lenses = Lens::all();
+
+            return $this->_indexView($dataTable, "admin");
+        } else {
+            abort(401);
+        }
+    }
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @param Array  $lenses An array with all lenses
+     * @param String $user   user for a normal user, admin for an admin
+     *
+     * @return \Illuminate\Http\Response
+     */
+    private function _indexView($dataTable, $user)
+    {
+        return $dataTable->with('user', $user)->render('layout.lens.view');
+    }
+
+    /**
+     * Display a listing of the resource in JSON format.
+     *
+     * @param int $id The id of the lens to return
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function getLensJson(int $id)
+    {
+        $lens = Lens::findOrFail($id);
+
+        return response($lens->jsonSerialize(), Response::HTTP_OK);
     }
 
     /**
@@ -38,18 +113,22 @@ class LensController extends Controller
      */
     public function store(Request $request)
     {
+        $request['observer_id'] = auth()->id();
+
         $validated = request()->validate(
             [
                 'observer_id' => 'required',
-                'name' => ['required', 'min:5'],
-                'factor' => 'required'
+                'name' => ['required', 'min:6'],
+                'factor' => ['required', 'numeric', 'min:0', 'max:10'],
             ]
         );
 
         Lens::create($validated);
 
+        flash()->success(_i('Lens "%s" created', $request->name));
+
         // View the page with all lenses for the user
-        return view('layout.lens.view');
+        return redirect('/lens');
     }
 
     /**
@@ -61,7 +140,7 @@ class LensController extends Controller
      */
     public function show(Lens $lens)
     {
-        //
+        return view('layout.lens.show', ['lens' => $lens]);
     }
 
     /**
@@ -73,6 +152,8 @@ class LensController extends Controller
      */
     public function edit(Lens $lens)
     {
+        $this->authorize('update', $lens);
+
         return view('layout.lens.create', ['lens' => $lens, 'update' => true]);
     }
 
@@ -86,25 +167,33 @@ class LensController extends Controller
      */
     public function update(Request $request, Lens $lens)
     {
+        $this->authorize('update', $lens);
+
+        $request['observer_id'] = $lens->observer_id;
+
         // If the factor is set, the name should also be set in the form.
         if ($request->has('factor')) {
             $validated = request()->validate(
                 [
                     'observer_id' => 'required',
-                    'name' => ['required', 'min:5'],
-                    'factor' => 'required'
+                    'name' => ['required', 'min:6'],
+                    'factor' => ['required', 'numeric', 'min:0', 'max:10'],
                 ]
             );
 
             $lens->update(['factor' => $request->get('factor')]);
             $lens->update(['name' => $request->get('name')]);
+
+            flash()->warning(_i('Lens "%s" updated', $lens->name));
         } else {
             // This is only reached when clicking the active checkbox in the
             // lens overview.
             if ($request->has('active')) {
                 $lens->active();
+                flash()->warning(_i('Lens "%s" is active', $lens->name));
             } else {
                 $lens->inactive();
+                flash()->warning(_i('Lens "%s" is not longer active', $lens->name));
             }
         }
 
@@ -120,8 +209,14 @@ class LensController extends Controller
      */
     public function destroy(Lens $lens)
     {
-        $lens->delete();
+        $this->authorize('update', $lens);
 
-        return view('layout.lens.view');
-    }
+        if ($lens->observations > 0) {
+            flash()->error(_i('Lens "%s" has observations. Impossible to delete.', $lens->name));
+        } else {
+            $lens->delete();
+
+            flash()->error(_i('Lens "%s" deleted', $lens->name));
+        }
+        return redirect()->back();    }
 }
